+++ addons/zylann.hterrain/hterrain.gd
@@ -157,6 +157,8 @@ var _shader_type := SHADER_CLASSIC4_LITE
 var _shader_uses_texture_array := false
 var _material := ShaderMaterial.new()
 var _material_params_need_update := false
+# Possible values are the same as the enum `GeometryInstance.SHADOW_CASTING_SETTING_*`.
+var _cast_shadow_setting := GeometryInstance.SHADOW_CASTING_SETTING_ON
 
 var _render_layer_mask := 1
 
@@ -322,6 +324,13 @@ func _get_property_list():
 			"type": TYPE_INT,
 			"usage": PROPERTY_USAGE_EDITOR | PROPERTY_USAGE_STORAGE,
 			"hint": PROPERTY_HINT_LAYERS_3D_RENDER
+		},
+		{
+			"name": "cast_shadow",
+			"type": TYPE_INT,
+			"usage": PROPERTY_USAGE_EDITOR | PROPERTY_USAGE_STORAGE,
+			"hint": PROPERTY_HINT_ENUM,
+			"hint_string": "Off,On,DoubleSided,ShadowsOnly"
 		}
 	]
 
@@ -382,6 +391,9 @@ func _get(key: String):
 	elif key == "render_layers":
 		return get_render_layer_mask()
 	
+	elif key == "cast_shadow":
+		return _cast_shadow_setting
+	
 
 func _set(key: String, value):
 	if key == "data_directory":
@@ -440,6 +452,9 @@ func _set(key: String, value):
 	elif key == "render_layers":
 		return set_render_layer_mask(value)
 
+	elif key == "cast_shadow":
+		set_cast_shadow(value)
+
 
 func get_texture_set() -> HTerrainTextureSet:
 	return _texture_set
@@ -481,6 +496,17 @@ func set_render_layer_mask(mask: int):
 
 func get_render_layer_mask() -> int:
 	return _render_layer_mask
+
+
+func set_cast_shadow(setting: int):
+	if setting == _cast_shadow_setting:
+		return
+	_cast_shadow_setting = setting
+	_for_all_chunks(HT_SetCastShadowSettingAction.new(setting))
+
+
+func get_cast_shadow() -> int:
+	return _cast_shadow_setting
 
 
 func _set_data_directory(dirpath: String):
@@ -1327,6 +1353,7 @@ func _cb_make_chunk(cpos_x: int, cpos_y: int, lod: int):
 		chunk.parent_transform_changed(get_internal_transform())
 
 		chunk.set_render_layer_mask(_render_layer_mask)
+		chunk.set_cast_shadow_setting(_cast_shadow_setting)
 
 		var grid = _chunks[lod]
 		var row = grid[cpos_y]
@@ -1605,3 +1632,10 @@ class HT_SetRenderLayerMaskAction:
 	func exec(chunk):
 		chunk.set_render_layer_mask(mask)
 
+
+class HT_SetCastShadowSettingAction:
+	var setting := 0
+	func _init(s: int):
+		setting = s
+	func exec(chunk):
+		chunk.set_cast_shadow_setting(setting)