+++ addons/zylann.hterrain/hterrain_chunk.gd
@@ -118,3 +118,8 @@ func set_render_layer_mask(mask: int):
 	assert(_mesh_instance != RID())
 	VisualServer.instance_set_layer_mask(_mesh_instance, mask)
 
+
+func set_cast_shadow_setting(setting: int):
+	assert(_mesh_instance != RID())
+	VisualServer.instance_geometry_set_cast_shadows_setting(_mesh_instance, setting)
+